Just An Idea
I wouldn't exactly just accept this. The idea is to extrapolate out each
blog post into it's corresponding title/article name and loop over each
one defined in the data.json file. This keeps the index itself super dry
along with keeping each blog post in a single file. Pull this down as a
separate branch, play with it, and likely just delete it.